fix: edit some file references after refactor
